Mark crate as un-maintained
I no longer work with `rust-libp2p` and do not wish to follow the
developments of the project, maintaining this crate in any meaningful
way is not possible without doing so. As such, I am ceasing to maintain
this crate.

If this crate is useful to you and you would like to take over
maintenance please email me: me@tobin.cc

Thanks,
Tobin.